LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerrawupdate_flush_pointercapture.https-expected.txt


FAIL Pointer capture state should be updated before `pointerrawupdate` assert_equals: expected "pointerrawupdate@target, pointermove@target, pointerdown@target, gotpointercapture@target, pointerrawupdate@target, pointermove@target, pointerup@target, lostpointercapture@target, pointerover@dest" but got "pointermove@target, pointerdown@target, gotpointercapture@target, pointermove@target, pointerup@target, lostpointercapture@target"
